# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Todo Microservice API
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
paths:
  /users/by-matrix-id:
    get:
      security:
        - bearerAuth: []
      summary: Get user by Matrix ID
      operationId: getUserByMatrixId
      parameters:
        - in: query
          name: matrixId
          schema:
            type: string
          required: true
          description: Matrix user ID
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /auth/matrix/openid:
    post:
      summary: Authenticate using Matrix OpenID
      operationId: PostMatrixAuth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatrixOpenIDRequest"
      responses:
        "200":
          description: Matrix authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatrixAuthResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized - Matrix token verification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      required:
        - id
        - email
        - matrix_id
      properties:
        id:
          type: string
          format: uuid
          description: ID of the user
        email:
          type: string
          format: email
          description: Email of the user
          example: "newuser@example.com"
        matrix_id:
          type: string
          description: Matrix ID of the user
          example: "@user:matrix.example.com"
        created_at:
          type: string
          format: date-time
          description: Timestamp when the user was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the user was last updated
...
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: "Something went wrong"
...
    MatrixOpenIDRequest:
      type: object
      required:
        - access_token
        - matrix_server_name
      properties:
        access_token:
          type: string
          description: Matrix OpenID access token
          example: "matrix_access_token_123"
        matrix_server_name:
          type: string
          description: Matrix homeserver name
          example: "matrix.example.com"
    MatrixAuthResponse:
      type: object
      required:
        - token
        - mxid
        - user_id
      properties:
        token:
          type: string
          description: JWT token for authentication
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        mxid:
          type: string
          description: Matrix user ID
          example: "@user:matrix.example.com"
        user_id:
          type: string
          format: uuid
          description: ID of the user in the todo service
          example: "123e4567-e89b-12d3-a456-426614174000"
...